All functionalities working, Changed from datastore to Objectify, All the styles were updated(except button hover:cursor)

+++ src/main/webapp/playlist-items.jsp
@@ -4,6 +4,7 @@
     <%@ page import="DTO.VideoDTO" %>
     <%  String userEmail = (String)request.getAttribute("userEmail");
     ArrayList<VideoDTO> playListItems = (ArrayList<VideoDTO>) request.getAttribute("playListItems");
+    if(playListItems == null) playListItems = new ArrayList<VideoDTO>();
     request.setAttribute("userEmail",userEmail);
     String playListName = (String)request.getAttribute("playListName");
     %>
@@ -23,23 +24,27 @@
 
         <div class="playlist-items-container">
             <div class="main-container">
+
                 <h3>Current Playlist : <%=playListName%></h3>
-                <table>
-                    <div class="java"><% for(VideoDTO video : playListItems) { %></div>
-                    <div>
-                        <tr>
-                           <td> <a href="<%=video.getLink()%>"><%=video.getTitle()%></a></td>
-                           <td> <form action="DeleteVideoFromPlayList" method="POST">
-                                <input type="hidden" value="<%=video.getTitle()%>" name="videoTitle">
-                                <input type="hidden" value="<%=userEmail%>" name="userEmail">
-                                <input type="hidden" value="<%=playListName%>" name="playListName">
-                                <button type="submit" >Delete From Playlist</button>
-                            </form>
-                            </td>
-                        </tr>
-                    </div
-                     <% } %>
-                </table>
+                    <% if(playListItems.size()==0) { %>
+                        <div><br><center><h3>No video Found In Playlist</h3><center><div>
+                        <% } %>
+                        <table>
+                            <div class="java"><% for(VideoDTO video : playListItems) { %></div>
+                            <div>
+                                <tr>
+                                   <td> <a href="<%=video.getLink()%>" target="_blank"><%=video.getTitle()%></a></td>
+                                   <td> <form action="DeleteVideoFromPlayList" method="POST">
+                                        <input type="hidden" value="<%=video.getTitle()%>" name="videoTitle">
+                                        <input type="hidden" value="<%=userEmail%>" name="userEmail">
+                                        <input type="hidden" value="<%=playListName%>" name="playListName">
+                                        <button type="submit" >Delete From Playlist</button>
+                                    </form>
+                                    </td>
+                                </tr>
+                            </div
+                             <% } %>
+                        </table>
             </div>
         </div>
     </body>
